[#88][#1055] Readline: Combining history from concurrent sessions didn't work & minor changes
Issue
-----
If you start two sessions a and then b, perform operations in a and b,
and then close b and then a, b's history gets added first, but then a's
history overwrites it. The history should be combined, so that the
session that exits first writes its history first and then the entry
that exits next writes it history next.

Fix
---
Keep track of number of entries added in the current history via global
variable `ydb_rl_entries_count` and using `append_history` instead of
`write_history` for that many entries at exit (but clamp to
`history-size` if necessary).

Misc changes
------------
1. Change `~/.inputrc` application name from `yottadb` to `YottaDB`. I
   noticed that all the examples in the Readline documentation were like
   "Bash".
2. I originally forgot in the main MR for #88
   (d51962f0ed906c31ec18c4b009a121bf8eab349d) to truncate the history
   file on disk. If we don't do this, the history file will keep
   expanding. It's now always truncated to 1000 entries.
3. An `#endif` for `sr_unix/readline.h` was in the wrong place.
   Apparently it didn't matter as the header file was not included more
   than once, but it should be fixed.