[DEBUG-ONLY] Add asserts that CURRSOCKOFR error is impossible
Background
----------
Below is pasted from https://gitlab.com/YottaDB/DB/YDBTest/-/issues/538#note_1791404905

**CURRSOCKOFR**

After going through the code for a bit, I came to the conclusion that this error is not possible to
be reached normally. If ever that code path is reached, it is an out-of-design situation. I do see an
`assert(FALSE)` in the code path in `iosocket_open.c` and `iosocket_use.c` but don't see it in other
places where `ERR_CURRSOCKOFR` error is issued. I suspect that is just a mistake and that all places
where it is issued should have an `assert(FALSE)` before it. I will take care of that coding change
separately. So this does not need to be tested.

Fix
---
* All code paths that issue a `ERR_CURRSOCKOFR` error are modified to have an `assert(FALSE);` before
  the error is issued. This way we will be alerted in case this ever happens in real life in a Debug
  build.