[#1068] Add CONVERT, HOSTSYNC, TTSYNC device parameters, add missing space in HUPENABLE and change output of NOCENABLE for ZSHOW "D"
Background
---

Below is pasted from #1068

I have carefully reviewed the YottaDB online manual for using terminal devices, here: https://docs.yottadb.com/ProgrammersGuide/ioproc.html#using-terminals

I have gone through each parameter and turned it on and off and noted how each parameter affects ZSHOW "D". Some parameters show up when on, some when off and some have other patters.

But there are 3 parameters that are not shown regardless: CONVERT, HOSTSYNC, TTSYNC, HUPENABLE

FEATURE REQUESTS:

* Please include output in ZSHOW "D" showing status of terminal device for CONVERT
* Please include output in ZSHOW "D" showing status of terminal device for HOSTSYNC
* Please include output in ZSHOW "D" showing status of terminal device for TTSYNC
* Please include output in ZSHOW "D" showing status of terminal device for HUPENABLE

Change
---

- Bugfix in `sr_unix/iott_use.c`, changes to `t.c_iflag` and `t.c_lflag` are not stored back in `tt_ptr` after calling `Tcsetattr` and verifying that it was successful. Now fixed.
- Bugfix in `sr_unix/zshow_devices.c` -- Add missing space after for `HUPENABLE`
  - Previously, parameter that come after `HUPENABLE` will be concatenated with it.
- Add TTSYNC in `sr_unix/zshow_devices.c`
  - This will output in `ZSHOW "D"` when enable as `TTSYNC` and when disable as `NOTTSYNC`
- Add HOSTSYNC in `sr_unix/zshow_devices.c`
  - This will output in `ZSHOW "D"` when enable as `HOSTSYNC` and when disable as `NOHOST`
- Add CONVERT in `sr_unix/zshow_devices.c`
  - This will output in `ZSHOW "D"` when enable as `CONVERT` and not show in ZSHOW "D" when disable like HUPENABLE which does the same.
- Change zshow_param_index in `sr_port/deviceparameters.c` for more appropriate output.
  - For TTSYNC : `TTSYNC` (No change)
  - For NOTTSYNC : `NOTTSY` \> `NOTTSYNC`
  - For HOSTSYNC : `HOS` \> `HOSTSYNC`
  - For NOHOSTSYNC : `NOHOST` to `NOHOSTSYNC`
  - For NOCENABLE : `NOCE` to `NOCENABLE` And for changes in `sr_port/deviceparameters.c`

```diff
diff --git a/sr_port/deviceparameters.c b/sr_port/deviceparameters.c
index 21857808..8c4de9ad 100644
--- a/sr_port/deviceparameters.c
+++ b/sr_port/deviceparameters.c
@@ -3,7 +3,7 @@
  * Copyright (c) 2001-2022 Fidelity National Information	*
  * Services, Inc. and/or its subsidiaries. All rights reserved.	*
  *								*
- * Copyright (c) 2022 YottaDB LLC and/or its subsidiaries.	*
+ * Copyright (c) 2022-2024 YottaDB LLC and/or its subsidiaries.	*
  * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
@@ -107,7 +107,7 @@ LITDEF nametabent dev_param_names[] =

 	,{3,"NEW*"}	,{10,"NEWVERSION"}
 	,{4,"NOCA*"}	,{11,"NOCANONICAL"}
-	,{4,"NOCE*"}	,{6,"NOCENA*"}		/*,{9,"NOCENABLE"}*/
+	,{4,"NOCE*"}	,{6,"NOCENA*"}		,{9,"NOCENABLE"}
 	,{6,"NOCONV*"}	,{9,"NOCONVERT"}
 	,{6,"NODELI*"}	,{11,"NODELIMITER"}
 	,{6,"NODEST*"}	,{9,"NODESTROY"}
@@ -121,7 +121,7 @@ LITDEF nametabent dev_param_names[] =
 	,{5,"NOFIX*"}	,{7,"NOFIXED"}
 	,{4,"NOFLA*"}	,{5,"NOFLAG"}
 	,{5,"NOFOL*"}	,{8,"NOFOLLOW"}
-	,{5,"NOHOS*"}	,{6,"NOHOST*"}		/*,{10,"NOHOSTSYNC"}*/
+	,{5,"NOHOS*"}	,{6,"NOHOST*"}		,{10,"NOHOSTSYNC"}
 	,{4,"NOHU*"}	,{11,"NOHUPENABLE"}
 	,{4,"NOIN*"}	,{6,"NOINSE*"}		/*,{8,"NOINSERT"}*/
 	,{4,"NOPAG*"}	,{5,"NOPAGE"}
@@ -246,7 +246,7 @@ LITDEF	uint4 dev_param_index[27] =
 	0,   5,   9,   27,  35,  50,  66,  68,  73,  82,  82,  83,  91,  94,

 /*	O    P    Q    R    S    T    U    V    W    X    Y    Z    end	     */
-	162, 171, 190, 191, 206, 226, 236, 242, 243, 258, 259, 260, 283
+	164, 173, 192, 193, 208, 228, 238, 244, 245, 260, 261, 263, 285
 };

 /* Offset of string within letter in dev_param_names */
@@ -257,11 +257,11 @@ LITDEF zshow_index zshow_param_index[] =
 /*	ALLO    BLOC     COMMAND CONV     CTRA     DELE    DEST    EBCDIC   EDIT     EMPTERM  EXCE     EXTE */
 	{0,0},  {2,1},   {9,2},  {14,2},  {16,2},  {0,3},  {4,3},  {1,4},   {4,4},   {7,4},   {12,4},  {14,4},
 /*	FIE     FIL     FIXED   FOLLOW   HOST    ICHSET  INDEPENDENT  INSE    LAB      LENG */
-	{4,5},  {7,5},  {9,5},  {15,5},  {1,7},  {0,8},  {3,8},	      {7,8},  {0,11},  {2,11},
+	{4,5},  {7,5},  {9,5},  {15,5},  {2,7},  {0,8},  {3,8},	      {7,8},  {0,11},  {2,11},
 /*	NOCENE   NODEST    NOECHO    NOEDIT    NOEMPTERM NOESCA    NOFOLLOW  NOHOST     NOINSE */
-	{4,13},  {10,13},  {15,13},  {17,13},  {19,13},  {21,13},  {30,13},  {32,13},  {36,13},
+	{6,13},  {11,13},  {16,13},  {18,13},  {20,13},  {22,13},  {31,13},  {34,13},  {38,13},
 /*	NOPAST   NOREADS  NOTTSY    NOTYPE   NOWRAP */
-	{39,13}, {42,13}, {57,13},  {59,13}, {64,13},
+	{41,13}, {44,13}, {60,13},  {61,13}, {66,13},
 /*	OCHSET   PAD      PARSE     PAST      PRMMBX    RCHK     READ     READS    REC */
 	{1,14},  {8,15},  {12,15},  {13,15},  {17,15},  {0,17},  {1,17},  {3,17},  {5,17},
 /*      SHAR     SHELL    STDERR    STREAM    TERM      TTSY     TYPE     UIC */
@@ -358,7 +358,7 @@ int deviceparameters(oprtype *c, char who_calls)

 		,iop_newversion ,iop_newversion
 		,iop_nocanonical ,iop_nocanonical
-		,iop_nocenable ,iop_nocenable
+		,iop_nocenable ,iop_nocenable ,iop_nocenable
 		,iop_noconvert ,iop_noconvert
 		,iop_nodelimiter ,iop_nodelimiter
 		,iop_nodestroy ,iop_nodestroy
@@ -372,7 +372,7 @@ int deviceparameters(oprtype *c, char who_calls)
 		,iop_nofixed ,iop_nofixed
 		,iop_noflag ,iop_noflag
 		,iop_nofollow ,iop_nofollow
-		,iop_nohostsync ,iop_nohostsync
+		,iop_nohostsync ,iop_nohostsync ,iop_nohostsync
 		,iop_nohupenable ,iop_nohupenable
 		,iop_noinsert ,iop_noinsert
 		,iop_page ,iop_page
```

First, for enable `NOCE` to `NOCENABLE` for `NOCENABLE` and `NOHOST` to `NOHOSTSYNC` for `NOHOSTSYNC` we need to uncomment ,{9,"NOCENABLE"} and {10,"NOHOSTSYNC"} in `dev_param_names` first. After that, we need to change `dev_param_index` to match `dev_param_names` according to comment below

```c
/* Offset of letter in dev_param_names.  Adding e.g. 1 entry there will add 1 to every entry corresponding to subsequent letters */
```
We added 2 entry in `dev_param_names` so we need to add 2 in all character after N which are from O to end in `dev_param_index`.
Then, we need to add `iop_nocenable` and `iop_nohostsync` in `dev_param_data` that in the same order as `dev_param_names`.

For changing `NOCE` to `NOCENABLE`, we need to change `zshow_param_index` from {4,13} to {6,13} because these 2 values are index for selecting an entry in `dev_param_names`. `13` means select section 13 that start counting from 0 (Which is section `N`), with 4 as a entry number that start counting from 0 (Which is `NOCE`) to `6,13` which in section 'N' and 58 represent for `NOCENABLE`.

For this change, because we added 1 entry we also need to change `zshow_param_index` for every entry after `NOCENABLE` to before `NOHOSTSYNC` to add 1 for first sub-index value in each index e.g. for `NODEST` changed from {10,13} to {11,13}.

For changing `NOHOST` to `NOHOSTSYNC` also need to change from {32,13} : section 'N' and entry number 34 (`NOHOST`) to {34,13} : section 'N' and entry number 2 (`NOHOSTSYNC`).

For this change we also need to change `zshow_param_index` from every entry after `NOHOSTSYNC` to `NOWRAP` (end of section 'N') we need to add 2 for first sub-index value in each index because we added to entry here e.g. for `NOINSE` changed from {36,13} to {38,13}.

For changing `NOTTSY` to `NOTTSYNC` need to change {57,13} to {60,13} because we already added 2 for first sub-index and we need to add more 1 to pointing to `NOTTSYNC`

And for changing `HOS` to `HOSTSYNC`, no need to add value here need because this handled by changed `dev_param_index` earlier to so we only need to change from {1,7} : section 'H' and entry number 1 (`HOS`) to {2,7} : section 'H' and entry number 2 (`HOSTSYNC`).