[Pipeline] Fix pipeline failure after merge of GT.M V7.0-004 in docker-create stage
Background
---

GT.M V7.0-004 changed CMakeLists.txt to require ssl.h. This cause docker-create stage in pipeline to fail
because packages that related to ssl.h is not available in Dockerfile. So, all of Dockerfiles need to
add packages that related to `ssl.h` to fix the pipeline issue.

Change
---

- Dockerfile : This dockerfile is using Ubuntu 22.04 as an OS, so we need to add libssl-dev, libgcrypt-dev and gawk
- Dockerfile-debian : This dockerfile is using Debian Bullseye as an OS, so we need to add libssl-dev, libgcrypt-dev and gawk (same as Ubuntu)
- Dockerfile-rocky : This dockerfile is using Rocky as an OS so the packages name will be a little different which are openssl-devel, libgcrypt-devel and gawk.