[#1071] Use the new pipeline set-up
This depends on YDBTest!2120.

The new pipeline runs in a single script after the build. The script
runs multiple tests concurrently, choosing tests and subtests at random.
The tests are run concurrently. This replaces the current set-up which
ran all subtests in a known lists of tests all the time. The reason we
replaced it is adding more and more tests would not scale well. So we
opted to run a random set of tests.

There will be a future commit to speed up the pipeline as it now takes 6
minutes since it needs to download the image. There is way to limit the
download to only when the image has changed, and we plan to do that in
the future.