Allow running the pre-commit hook without a global install of YDB
In some cases, particularly when `sr_port/*errors.msg` is modified, the
precommit hooks runs yottadb to generate source code. Previously, it required
having either gtm_dist set in the environment, or a global install of YDB.

This changes the hook to default to using the newly built `build/
yottadb` in the current directory if it cannot find a global
installation.