[#873] Implement ZSHOW "V"::$STACK
When creating a new symbol table, save the current value of $STACK. When looking up variables for ZSHOW, iterate back over the symbol tables until we find one whose $STACK is <= the desired level, then use it for printing.

- Support the new stack level for each of the output forms (locals, globals, indirection, device output)
- Use -1 as an alias for $STACK, to simplify the code for existing calls to op_zshow
- Add a new STACKRANGE error and emit it if the stack level is negative or greater than $STACK.
- Extend op_fnlvname to take the symbol table as a parameter, rather than hard-coding the current symbol table. Update the relevant opcodes.
- Take care in `gtm_ztrigger` to increment the stack level when creating a new stack frame.

Note that the code in `m_zshow` for generating triples is somewhat hard to follow because it's interspersed with parsing.
See `op_indrzshow` for a better idea of how the new stack level is represented in the opcode triples.