[#665] Optimize consecutive global value accesses to naked references
This implements bullets 1 and 6 of https://gitlab.com/YottaDB/DB/YDB/-/issues/665#note_2326164647.

The basic idea is that for each opcode, do the following:
- GVNAME (if this variable does not match the previous access): record the variable, and do not do any optimization.
- GVNAME (if this variable matches the previous access): optimize to a new GVNAMENAKED opcode. See below for details.
- GVNAKED (if the subscripts do not match the previous access): reset the previously accessed global name
- Opcodes known to affect control flow (e.g. XECUTE) or modify $REFERENCE (e.g. TROLLBACK): reset previously accessed name.
- Any other opcode: do nothing.

Additionally, this sets a runtime flag on interrupts, function calls, or GOTO which instructs the runtime to skip the optimization.

More formally, we optimize when ALL of the following are true:
- The naked reference must be syntactically legal (e.g. `^X` will never be optimized)
- All subscripts except the last are M literal expressions.
- At compile time, for each opcode between the previous and current global access, either:
  - GVNAME: The new name must have the same or more number of subscripts as the old one.
  - GVNAKED: All accesses must have exactly 1 subscript.
  - NEWINTRINSIC or SVPUT: The intrinsic variable being modified is not $ZGBLDIR.
  - All other opcodes: The opcode cannot possibly modify $REFERENCE. For
    example, INDSAVGLVN always disables the optimization, even if it
    doesn't modify $REFERENCE at runtime, because we don't verify that
    at compile time.
- At runtime, between the previous and current global access, NONE of the following occur:
  - a function call (e.g. `op_callw`)
  - GOTO
  - ZSTEP
  - ZBREAK
  - $ZSTEP executes
  - $ZINTERRUPT executes
  - $ETRAP or $ZTRAP executes
  - An EXCEPTION handler for device open executes

---

About GVNAMENAKED: at runtime, we may execute code in an unexpected order due to
an interrupt or GOTO-with-offset. Set a runtime flag checked by the new GVNAMENAKED opcode.
That flag disables the optimization until the next global reference, at
which point we are back in a known state.

To support this, GVNAMENAKED takes all the parameters of a GVNAME, even
if it only ends up doing a GVNAKED at runtime.

Additionally, bump the object file format version to account for
the new OC_GVNAMENAKED.

-------

This adds a lot of new control-flow analysis, which changes the behavior
of M programs in strange ways if it is incorrect. To make it easier to
debug, GVNAMENAKED verifies `gv_optimize` at runtime in DBG builds by
cross-checking GVNAKED against GVNAME.

Here is an example error message:
```
%YDB-E-GVDBGNAKEDMISMATCH, Invalid GVNAKED in gv_optimize: $REFERENCE did not match OP_GVNAKED: ^bxxx(2) != ^cxxx(2). Opcodes seen: OC_NOOP|OC_PARAMETER|OC_LIT|OC_GVNAME|OC_GVPUT|OC_SVGET|OC_JMP|OC_JMPEQU|OC_LINESTART|OC_ILIT|OC_GVZWITHDRAW|OC_BOOLEXPRSTART|OC_BOOLEXPRFINISH|OC_EQU_RETBOOL|
                At M source location +1^cxxx#1#
```

In that case, the error was that OC_JMP and OC_JMPEQU should have been
marked as resetting $REFERENCE.

This check only happens in DBG builds. In PRO builds we only do a
GVNAKED (when not illegal due to an interrupt).

-------

Fix SIG-11 when issuing GVDBGNAKED error if gtmcompile=-dynamic_literals

* See https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1636#note_2445477594 for how to
  reproduce the SIG-11.