[#1127] Mark libyotttadbutil.so as not needing executable stack (avoids DLLNOOPEN error in glibc 2.41)
YottaDB does not run with glibc 2.41, and the problem is that `libyottadbutil.so` is marked as having an executable stack. Since it does not need that capability, it should be be built without an executable stack.

```
$ /usr/local/lib/yottadb/r203/yottadb -run %XCMD 'w $zyre,!'
%YDB-E-DLLNOOPEN, Failed to load external dynamic library /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
%YDB-I-TEXT, /extra/usr/local/lib/yottadb/r203/libyottadbutil.so: cannot enable executable stack as shared object requires: Invalid
argument
$ execstack -q /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
X /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
$
```

If the stack is marked as not executable it runs.

```
$ sudo patchelf --clear-execstack /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
$ sudo execstack -q /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
- /extra/usr/local/lib/yottadb/r203/libyottadbutil.so
$ /usr/local/lib/yottadb/r203/yottadb -run %XCMD 'w $zyre,!'
YottaDB r2.03 Linux x86_64
$
```

Instead of patching a built shared library, the `ld` flag `-z noexecstack` flags the binary as not having an executable stack when it is being built.