[#1133] Support `-machine` in release mode
We already supported this in debug mode, so mostly this just removes
`#ifdef DEBUG` macros. Additionally, this makes `-machine` imply `-list`.

Fix a pre-existing assertion error in `list_file` when the source file
name was longer than `MAX_MIDENT_LEN`.

Finally, remove unused `generated_details` buffers.

---

Another pre-existing bug fixed here is that input files were assumed to
end in `.m`, and YDB behaved poorly when that wasn't the case.
In particular they were truncated to "two characters shorter than the
actual length", creating the wrong file when the name was at least 2
characters and causing memory corruption when it was only 1 character.

Fix both these issues by only truncating after checking the file does in
fact end with `.m`.

---

While working on this, clang-tidy gave several warnings of the following
form:
```
> emit_code.c:warning: the result from calling 'memcpy' is not null-terminated [bugprone-not-null-terminated-result]
```

I inspected these and all of them were because while emitting code, code_buf is not nul-terminated. For example, this line gave one of the warnings:

```
/* 1246 */ memcpy(obpt, &vdat_def[0], VDAT_DEF_SIZE);
/* 1247 */ obpt += VDAT_DEF_SIZE;
```

Switch from `memcpy` to `stpcpy`, which is less error-prone than
requiring the sizes to be in sync. Additionally remove VDAT_* constants
now that they are no longer needed.

`stpcpy` requires its strings to be nul-terminated, so change the
declaration of `vxi_opcode` to be long enough to include a
nul-terminator.

While working on *that*, a secondary bug occurred: the length of
`vxi_opcode` did not match between `emit_code.c` and `mtables.c`.
And that caused a bug where indexing into the array did not give the
right string literal back. Avoid a reoccurence of this bug by moving the
declaration of `vxi_opcode` into `vxi.h` so it can't get out of sync.