[#1125] Appropriate permissions for /tmp/yottadb and group for /tmp/yottadb/<rel>
When sourcing `ydb_env_set`, %YDBENV creates `/tmp/yottadb` and `/tmp/yottadb/$ydb_rel` if they don't exist. `/tmp/yottadb` needs to be world writable and executable because different users may source `ydb_env_set` from different releases, and each new $ydb_rel subdirectory will be created by the user first using that release. To prevent a user from modifying a $ydb_rel subdirectory created by a different user, `/tmp/yottadb` has the sticky bit set with this merge request. Previously it did not have the sticky bit set.

Since YottaDB can be installed with a group restriction (e.g., `--group xyz --group-restriction`), then only members of that group, or root, have access to `/tmp/yottadb/$ydb_rel`. However, if a root process first sources `ydb_env_set` (i.e., executes %YDBENV), then the directory would previously have root as the user and group. This meant that a normal user would not subsequently be able to create files in that directory. With this merge request, if YottaDB is installed with a group restriction, the group of the $ydb_rel subdirectory is set to the group to which execution is restricted. (If execution of YottaDB is not restricted, then the $ydb_dir subdirectory is world writable, so it is OK for the user and group to be root.)

A redudant section of code has been removed, which dealt with the edge case where YottaDB was installed owned by a user (e.g., `--user ydbuser`) who was not a member of a group to which YottaDB execution was restricted. This edge case (where YottaDB execution is restricted for members of a group as well as a specific user who is not a member the group) cannot occur with YottaDB because although GT.M's `configure` permits such an installation, YottaDB's `ydbinstall` does not:

```
$ sudo -E ./ydbinstall --utf8 --allplugins --overwrite-existing --group-restriction --group ydb --user endeavouros
endeavouros is not a member of ydb
YottaDB installation aborted due to above error. Run ydbinstall --help for detailed option list
$
```

Furthermore, `ydb_env_set` and %YDBENV are specific to YottaDB. So code in %YDBENV to handle that upstream edge case can never be executed, and was removed.

It's interesting that the explanation of the changes in this merge request significantly exceeds the merge request itself!