[#474] Add SimpleAPI functions to support JSON
Finish the work started in commit 13234ff3 to add support for two new
SimpleAPI functions, and their SimpleThreadAPI wrappers. They add new
functionality that supports decoding a JSON string in to an M local or
global array tree or subtree using the functions
`ydb_decode_s()/ydb_decode_st()`, as well as encoding an M local or
global array tree or subtree in to a JSON string using the functions
`ydb_encode_s()/ydb_encode_st()`.

There are some limitations, because JSON and M arrays don't map
perfectly between each other. M can store data in an array node that
also contains child subscripts, but JSON doesn't have a way to structure
that intuitively. This code works around that by storing the data at
such a node using an empty key in the JSON inside the object that
represents the child subscripts of that node, because such a key is
legal in JSON, but isn't likely to ever be used. YDB can be configured
to allow NULL subscripts, but is not by default, and probably not in
most configurations, thus making the mapping to a node with child
subscripts work decently enough.

Another limitation is that JSON supports three data types as values,
that M does not; namely `true`, `false`, and `null`. There is no way to
distinguish between those data types, and their string equivalents,
without adding semantic information nodes to the M array tree, and this
code does not do that at this time. Instead it maps those data types to
all upper-case string versions in the M array - "TRUE", "FALSE", and
"NULL" - which does prevent them from being encoded in to JSON as
strings, since they will be encoded as the JSON data types. JSON
document strings that contain both forms of one, or more, of these data
types are probably not very likely.

Another limitation is that YDB's implementation of real numbers are
fixed-point, with high precision within their supported range, while the
JSON library this code uses (Jansson), implements real numbers with
double precision floating-point (IEEE 754), which have a higher
supported range, but cannot represent every possible number within that
range. This can lead to slight changes to real numbers when moving
between the two. Note that JSON itself doesn't specify a format for real
numbers, nor does it even distinguish between a real number or an
integer.