[#1147] Create Debug/ASAN YottaDB on Ubuntu Docker Image
Dockerfile now passes ENABLE_ASAN down to CMake, which takes care of
building with ASAN. Note that `ENABLE_ASAN` needs to be specified for
each stage since each stage is actually a separate environment. See
https://docs.docker.com/build/building/variables/#scoping. Initially I
used `ENABLE_ASAN` with a default value of "OFF", but this will cause
`ASAN_OPTIONS` to get initialized; so now there is no default... this
way, `ASAN_OPTIONS` is empty if `ENABLE_ASAN` is not specified.

`.gitlab-ci.yml` adds a new image to Docker Hub called "yottadb-asan"
and "yottadb-asan-base".