[#1105] Pipeline: Run tests directly, not using dind
YDB pipeline previous ran using dind (as a docker run) (i.e. the Gitlab
runner ran a dind image, and then ran a docker run inside the dind
image). Now, we do the "docker run" internally in the Gitlab runner. The
advantage to do doing this is that we don't have to download the image
each time, which is what previously happened.

The reason it was done using the previous way is that it was easy to
reproduce problems locally as it was an identical set-up; but the
YDBTest image now is nearly 10 GB, so doing the best we can to minimize
downloads is best.

This change is actually done in an MR in YDBTest; the only change done
here is to change the `.gitlab-ci.yml` file.

Doing this will also help us reuse code for YDBTest#807, as right now the
different running methodology prevents us from re-using code between the
YDB and YDBTest pipelines.

Now tests take 4 minutes to run down from around 10 minutes.