[#1034] Fix MUPIP EXTEND to accept case-insensitive region names
After a thorough investigation, it was determined that the only MUPIP
operation that did not support case-insensitive region names was MUPIP
EXTEND. The investigation involved consulting the YDB documentation on
MUPIP, and running every operation that involves region names as a
possible option, in the gdb debugger, to determine how it handles them.

After that, the test suite was searched for all tests that involved the
above mentioned operations that involve region names, to ensure that
every one of them had at least one test that used mixed-case region
names. If an operation was found that did not have such a test case, an
existing one was modified. In every case, it was found that there was
always more than one test case using a region name, so in the end, no
tests lost any coverage, and coverage for case-insensitive names was
added.

Then, the test suite for every applicable test was run and that
confirmed that no other MUPIP operation was lacking case-insensitive
region name support.