[YDB#1018] Fix DBUPGRDREQ error from MUPIP UPGRADE after a JNLFNF/JNLSWITCHFAIL error
Background
----------
See discussion in below threads for more background on the issues noticed.

- [ ] https://gitlab.com/YottaDB/DB/YDB/-/issues/1018#note_2695983616
- [ ] https://gitlab.com/YottaDB/DB/YDB/-/issues/1018#note_2696001560

Fix
---
As mentioned in one of the above discussion threads, the fix is for mupip upgrade to move on
to the next region if it encounters an error opening the .mjl file (previously it used to
continue upgrading the .dat file even if there are errors opening the .mjl file).

And that is exactly what is implemented in this commit. We check the return from `jnl_ensure_open()`
and if it is non-zero, we abort the upgrade in this region, issue a message saying we are moving
on to the next region and actually move on to the next region (just like we do already for various
other error code paths).