[#876] Added `$ZGETJPI` keyword "CMDLINE"
Added a new keyword to the `$ZGETJPI` function. With the keyword `"CMDLINE"`, it will now return the command line of the process.

If the file does not exist or the program does not have permission to access the information, an empty string is returned. If the command line is longer than 1024 chars, `$ZGETJPI` returns only the first 1024 chars.