[COSMETIC] Fix a number of http URLs in source code comments to instead be https
A number of URLs in the source code were found to be http URLs rather than https URLs. Along the way, some content had migrated from one URL to another. This commit replaces http URLs with https URLs. The URLs that remain http URLs are:

- Those which have not migrated to https URLs.
- Those which no longer exist and for whose content a migrated URL was not found.

All https URLs in the commit have been validated. `ydbinstall` was changed as it was checking for both http and https links on SourceForge for the upstream code. This shell OR operation was removed and the test is now exclusively for the https link. A typo in LICENSE was fixed.