[#1176] Assert Fail while getting supplementary groups
When getting supplementary groups, an assert in sr_unix/gtm_permissions.c that was
designed to ensure that you initialize the supplementary group list before
 trying to access its data was failing.
The assert assumes that the supplementary group list is not size 0.
This commit ensures that the assert will not cause an error in the case
of 0 supplementary groups.

see more about the assert here https://gitlab.com/YottaDB/DB/YDB/-/issues/1176#note_2761328839

The assert will still trigger if the group list is uninitialized.