[#1152] M commands to serialize/deserialize local or global variable subtree
This feature adds support for two new commands, `ZYDE[CODE]` and
`ZYEN[CODE]`. They expose the functionality of the SimpleAPI functions,
`ydb_decode_s()` and `ydb_encode_s()`, to the M runtime. `ZYDECODE` decodes
JSON strings in to M arrays, and `ZYENCODE` encodes M arrays in to JSON
strings. They use the same syntax as the `MERGE` command, but use very
different semantics, as will be demonstrated below. They both support locals
and globals as their source or destination variables, as well as
unsubscripted or subscripted variables, including all forms of indirection
and subscript expressions. They affect the naked indicator and `$REFERENCE`
using the same semantics as `MERGE`. They also support full multi-node
chunking, using a specific format. This allows them to support JSON strings
larger than 1 MiB (the YottaDB string size limit) as both input to
`ZYDECODE` and output from `ZYENCODE`. The following examples show each
command in action.

The input to `ZYDECODE` is a JSON string, broken in to 1 MiB chunks. So if
the string is smaller than 1 MiB, you'd normally store it in one node. The
root of the source variable must contain a positive integer representing the
number of nodes containing the JSON string, and each chunk of the string,
must be stored in successive integers starting with 1.

Examples:

The following `ZYDECODE` example shows the input JSON string split in to two
chunks, to show that input chunking works. Normally, you'd put the whole
example string in `f(1)` and set `f` to `1` rather than `2`.

```
>set f(1)="{""array"": [1, 2, 3, 4, 5], ""key"": ""value"", ""null"": "
>set f(2)="null, ""string"": ""null""}"
>set f=2
>zydecode t("json")=f
>zwrite t
t("json","array",0)=1
t("json","array",1)=2
t("json","array",2)=3
t("json","array",3)=4
t("json","array",4)=5
t("json","key")="value"
t("json","null")=$C(0)_"null"
t("json","string")="null"
```

The following `ZYENCODE` example shows that even if the input JSON from a
prior `ZYDECODE` command was chunked unnecessarily, the output JSON of the
`ZYENCODE` command will only chunk once a string is longer than 1 MiB (or
the max record size if the destination is a global). The JSON output format
returned from `ZYENCODE` matches the JSON input format required by the
`ZYDECODE` command.

```
>zyencode json=t("json")
>zwrite json
json=1
json(1)="{""array"": {""0"": 1, ""1"": 2, ""2"": 3, ""3"": 4, ""4"": 5}, ""key"": ""value"", ""null"": null, ""string"": ""null""}"
```

New Errors:

There are a few runtime errors that can occur when input is not in the right
format, or if there is an error that has come back from one of the SimpleAPI
functions. The following are the new error mnemonics added for these
commands and their code numbers and text strings.

```
ERR_ZYDECODEDESC     151028202 ZYDECODE operation not possible: !AD is a descendant of !AD
ERR_ZYDECODEINCOMPL  151028208 Error encountered during ZYDECODE; operation may be incomplete
ERR_ZYDECODEWRONGCNT 151028218 Error encountered during ZYDECODE; JSON chunk count in root of source variable must be a positive integer
ERR_ZYENCODEDESC     151028226 ZYENCODE operation not possible: !AD is a descendant of !AD
ERR_ZYENCODEINCOMPL  151028232 Error encountered during ZYENCODE; operation may be incomplete
ERR_ZYENCODESRCUNDEF 151028242 Error encountered during ZYENCODE; source variable must not be undefined
ERR_BADZYENZYDENEST  151028250 Unsupported nesting of ZYENCODE or ZYDECODE command
```

There was also a change to an error that is sometimes returned by
`ydb_encode_s()`. The `ERR_INVSTRLEN` error was replaced by
`ERR_JANSSONINVSTRLEN`. This was done because the `ERR_INVSTRLEN` error
returns the length of the string that was too long, using a function that
works with the `unsigned int` type, so any value over `4294967295` bytes
(`UINT_MAX`), was truncated. The new `ERR_JANSSONINVSTRLEN` error is the
same error, but it uses a format and function that works with the `uint64_t`
type instead, so avoids truncation with large JSON strings.

```
ERR_JANSSONINVSTRLEN 151028194 Invalid string length !UJ: max !UJ
```

Bug Fixes:

There was also a change to the `ERR_MERGEDESC` error string format, to make
local variable references consistent with global variable references, such
that both use the `ZWRITE` format. There was also an improvement to the
grammar of the `ERR_MERGEDESC` error string. There was a fix to a bug in
processing the `ERR_MERGEDESC` error. The local symbol table for command
arguments might not contain the local variable reference for the actual
variable after resolving indirection, which could result in a missing local
variable reference.

This was fixed by using a new macro that was added called
`FIND_INDIRECT_SYMBOL_NAME` which scans the full symbol table to find the
variable's name, when it can't be located in the local symbol table for the
command argument frame.

When the source of a `ZYENCODE` command is a local variable that has a
single node, and that node's last subscript is a null subscript (or empty
string), and it is the only node in the local variable, the
`yed_encode_tree()` function, called by the `ydb_encode_s()` function could
end up in an infinite loop. This happens because the `ydb_node_next_s()`
function that it calls to get the next node in the variable, did not return
a `YDB_ERR_NODEEND` status to indicate the end of the variable tree, when
the last subscript is an empty string in a variable with a single node (when
allowed by setting the `ydb_lvnullsubs` environment variable to `1`).

This was fixed by adding an inline function that is called at the beginning
of the loop that orders through the nodes at the current subscript level, to
detect when the previous node, and the current node are identical, since
that can only happen in this one scenario.

Implementation Notes:

This feature required a bump to the `OBJ_UNIX_LABEL` version number due to
the additon of the following new opcodes.

```
OC_ZYDECODE      OC_ZYDECODE_GVARG  OC_ZYDECODE_LVARG  OC_INDZYDECODE
OC_INDZYDECODE2  OC_ZYENCODE        OC_ZYENCODE_GVARG  OC_ZYENCODE_LVARG
OC_INDZYENCODE   OC_INDZYENCODE2
```

There were a few changes to the `ydb_decode_s()` and `ydb_encode_s()` APIs.
They both contained some functions with unnecessary parameters, which were
removed. They also had a few duplicated error conditionals, which were also
removed. There were other minor changes to both APIs and the
`ydb_simpleapi_ch()` condition handler to make them work with the new
commands properly. There were also some memory usage optimizations.

One of the changes to `ydb_encode_s()` allows a `ydb_buffer_t` to be passed
without being preallocated, which allowed `ZYENCODE` to run twice as fast.
This was coupled with use of a different Jansson API `json_dumps()` that
doesn't require passing a preallocated buffer.