[#1142] Fix %YDB-E-SYSCALL error during compilation if more than one FALLINTOFLST warning
Background:
Issue link https://gitlab.com/YottaDB/DB/YDB/-/issues/1142
The original issue request.
If you have multiple FALLINTOFLST warnings in your file the compiler will fail
by expecting the wrong size of file. This was because the graph of commands
 was not properly linked together in line.c. It will now link the tree together.

To explain this I will first set out some vocabulary:

Each line has a parent, child, and sibling.
 Sibling is the next line at the same indentation level.
parent is the last line at the previous indentation level.
child is only set if the line is the last line before the next indentation level deeper.
If child is set, it is the first line of the next indentation level.

consider the following M program

 Do
  . indented Command
NextLabel(x)

Lets give shorthands for the lines in this program.
Label line or current line is the line that says NextLabel(x)
Previous line is the line that says. indented Command
Previous parent is the line that says Do,
Added quit is the implied line with a quit added by the compiler before the label line

Context:

Note that as the label is never indented, the Previous parent
would be the most recent non-indented line.
Currently, the code will never add a quit after
a non-indented line.

The compiler does not set the sibling of current line until it passes the next line
through the line.c function as the current line's sibling does not exist until then.
If current line is less indented than the previous line, the compiler follows
the chain of parents backwards until it reaches the Previous parent on
the same indentation level as current line, then that Previous
parent's sibling is set to current line.

Previous behavior:
Before this change, the compiler had a pointer that pointed to previous line called mline_tail.
It would then find mline_tail's prev parent on the same indentation level
of the current line(possibly mline_tail itself if it does not change indentation levels).
It would then set that line's sibling to be the current line.
The problem came when a return was added.
The mline_tail pointer would point at Added quit correctly setting added quit's sibling
to be current line, but meaning that prev parent's sibling was never set.
The compiler uses this graph to calculate the expected size of the output file.
When the error in the graph causes the expected size of the file
to be different from its actual size, the compiler exits with an error.

Fix:
The new behavior is that, when the Added quit is created, it automatically has its sibling set
to be the label line. Later in the function when the function normally sets the sibling of
the previous parent to be the current line, it will instead set the sibling
of the previous parent to be the added quit.

Additional fix:

This commit also fixes an internal bug, the line number of an Added quit,
added_ret->line_number,  was being recorded as the line after the label.
This was causing the tree(arrows pointing to siblings) to look like
line 1(prev parent)->line 4(added_ret)->line 3(lable line)->line 4(next line after label)
As the added return is executed before the label line, it should not be on a line after label line.
This matchmatch was happening as lnc, the variable pointing to the current line number,
is incremented at the top of the file right after being used to set the current line's line number.
The line number of the Added quit will now match the line of the label.
As best I can tell, the code currently only uses the line_number field when processing labels,
so this change has no effect currently but it will help people correctly understand the
information in a debugger and may prevent future bugs.

Dot_count change

In the case where you have a formal parameter line in the middle of an indented block,
that will produce an error. The previous behavior was that
the indentation level would return to what it was on the previous line, as suggested by the dots.
This has been changed to be indentation level 0, as suggested by the label.
This allows us to assert that siblings have the same parent. More information at discussion
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1773#note_2906349061
```
Do  w "hi,!"
 . w "indented",!
LABEL(x) . w "line",!
```